Swaps out custom ColorForm for one built from builder
This simplifies the Eye and Hair color forms to just use the
ColorForBuilder. It may be that we have other areas in the app where the
color specific stuff will want to change, and we just have a select
button, but for now, this is a good simplification that eliminates a lot
of duplication.
